# By Tyler Benfield (4) and others
# Via Wade Ashby
* 'master' of https://github.com/amlozano1/project3:
  Updated readme.txt
  Added compile and run instructions
  Fixed output to say response time for all situations.
  Added simulation response time.
  Added response time println in nanoseconds.
  Added memory analytics
  Fixed default precision value
  Removed unused code
  Timers in simulation

+++ src/base/PersistenceManager.java
@@ -38,7 +38,7 @@ public class PersistenceManager {
 	public void saveCells(int simulationId, int simulationResultId, Cell[][] data) {
 		int cellsSaved = 0;
 		int totalCells = 0;
-		short decimalPlaces = ((InvocationParms)System.getProperties().get(Utils.INVOCATION_PARAMETERS_KEY)).precision;
+		int decimalPlaces = ((InvocationParms)System.getProperties().get(Utils.INVOCATION_PARAMETERS_KEY)).precision;
 		for (int x = 0; x < data.length; x++) {
 			for (int y = 0; y < data[x].length; y++) {
 				float ratio = (float)cellsSaved/(float)(++totalCells);
